+++ src/components/CodeEditorWindow/Menu/CommandsDrawer/CommandsDrawer.jsx
@@ -7,19 +7,7 @@ import Close from '@mui/icons-material/Close';
 import useMediaQuery from '@mui/material/useMediaQuery';
 import { useSnackbar } from 'notistack';
 import { getSnackbarOptions } from '../../../Common/utils/snackbarOptions';
-
-const resolveRequiredParams = function (openapi, requestBodyContent) {
-  const contentType = Object.keys(requestBodyContent)[0];
-  if ('$ref' in requestBodyContent[contentType].schema) {
-    const refPath = requestBodyContent[contentType].schema.$ref;
-    // parse and navigate to the ref
-    const pathComponents = refPath.slice(2).split('/');
-    const schema = pathComponents.reduce((doc, pathComponent) => doc[pathComponent], openapi);
-    return schema.required;
-  } else {
-    return requestBodyContent[contentType].schema.required;
-  }
-};
+import { resolveRequiredBodyParams } from '../../config/CommandsDrawerUtils';
 
 const CommandsDrawer = ({ open, toggleDrawer, handleInsertCommand }) => {
   const [allCommands, setAllCommands] = useState([]);
@@ -41,7 +29,7 @@ const CommandsDrawer = ({ open, toggleDrawer, handleInsertCommand }) => {
               const hasRequestBody = !!data.paths[path][method].requestBody;
               let requiredBodyParameters = null;
               if (hasRequestBody) {
-                requiredBodyParameters = resolveRequiredParams(data, data.paths[path][method].requestBody.content);
+                requiredBodyParameters = resolveRequiredBodyParams(data, data.paths[path][method].requestBody.content);
               }
 
               return {
